Django-Middleware-0x03/requirements.txt
aiosqlite==0.21.0
annotated-types==0.6.0
appdata==2.2.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
colored==2.2.4
comtypes==1.2.0
cowsay==6.1
crispy-bootstrap4==2024.10
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.7
django-bootstrap-pagination==1.7.1
django-crispy-forms==2.3
django-debug-toolbar==4.3.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.2
emoji==2.12.1
Faker==37.4.2
filelock==3.13.1
fonttools==4.44.0
fpdf2==2.7.6
idna==3.4
inflect==7.0.0
iniconfig==2.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
Markdown==3.5.1
mypy==1.6.1
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
packaging==23.2
parameterized==0.9.0
Pillow==10.1.0
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.3.0
pycodestyle==2.14.0
pydantic==2.4.2
pydantic_core==2.10.1
pyfiglet==1.0.2
PyJWT==2.9.0
pypiwin32==223
pytest==7.4.2
python-decouple==3.8
python-dotenv==1.1.1
pyttsx3==2.90
pywin32==306
referencing==0.30.2
reportlab==4.0.8
requests==2.31.0
rpds-py==0.10.6
setuptools==69.1.1
sqlparse==0.4.4
tabulate==0.9.0
to-requirements.txt==2.0.11
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
validator-collection==1.5.0
validators==0.22.0
virtualenv==20.25.1
